style: format storage modules
